+++ app/src/main/java/com/techcourse/DispatcherServletInitializer.java
@@ -4,10 +4,11 @@ import jakarta.servlet.ServletContext;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import web.org.springframework.web.WebApplicationInitializer;
+import webmvc.org.springframework.web.servlet.mvc.tobe.DispatcherServlet;
 
 /**
  * Base class for {@link WebApplicationInitializer}
- * implementations that register a {@link DispatcherServlet} in the servlet context.
+ * implementations that register a {@link webmvc.org.springframework.web.servlet.mvc.tobe.DispatcherServlet} in the servlet context.
  */
 public class DispatcherServletInitializer implements WebApplicationInitializer {
