+++ mvc/src/main/java/webmvc/org/springframework/web/servlet/mvc/tobe/HandlerMappingRegistry.java
@@ -1,4 +1,4 @@
-package com.techcourse;
+package webmvc.org.springframework.web.servlet.mvc.tobe;
 
 import jakarta.servlet.http.HttpServletRequest;
 import webmvc.org.springframework.web.servlet.mvc.asis.HandlerMapping;
